backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
nltk==3.9.1
prisma==0.15.0
deep-translator==1.11.4
